Move test to `TestTorch`
